+++ test/fixtures/cookbooks/consul_spec/metadata.rb
@@ -1,4 +1,0 @@
-name 'consul_spec'
-version '0.1.0'
-
-depends 'consul'